// libs/authentication/authentication-core/src/lib/domain/models/AuthenticationErrorMessage.ts
export enum DefaultDomainErrorMessages {
  BAD_CREDENTIALS = 'BAD_CREDENTIALS',
  AUTH_INIT_REDIRECT_TO_ERROR_PAGE = 'AUTH_INIT_REDIRECT_TO_ERROR_PAGE',
  AUTH_INIT_NO_REDIRECT = 'AUTH_INIT_NO_REDIRECT',
  AUTH_INIT_SAML_REQUEST_FAIL = 'AUTH_INIT_SAML_REQUEST_FAIL',
  AUTH_INIT_TOO_MANY_ACTIVE_AUTH_SESSIONS_FOR_USER = 'AUTH_INIT_TOO_MANY_ACTIVE_AUTH_SESSIONS_FOR_USER',
  AUTH_INIT_UNKNOWN_USER = 'AUTH_INIT_UNKNOWN_USER',
  AUTH_INIT_UNKNOWN_ICG_VALIDATION_METHOD_TYPE = 'AUTH_INIT_UNKNOWN_ICG_VALIDATION_METHOD_TYPE',
  AUTH_INIT_FOLLOW_REDIRECT_FAIL = 'AUTH_INIT_FOLLOW_REDIRECT_FAIL',
  AUTH_INIT_AUTHENTICATION_LOCKED = 'AUTH_INIT_AUTHENTICATION_LOCKED',
  AUTH_INIT_AUTHENTICATION_FAILED = 'AUTH_INIT_AUTHENTICATION_FAILED',
  AUTH_INIT_AUTHENTICATION_FAILED_SMS_PROVIDER_ERROR = 'AUTH_INIT_AUTHENTICATION_FAILED_SMS_PROVIDER_ERROR',
  AUTH_VERIFY_FAIL = 'AUTH_VERIFY_FAIL',
  AUTH_VERIFY_MALFORMED_OTP_SMS_RESPONSE_BODY = 'AUTH_VERIFY_MALFORMED_OTP_SMS_RESPONSE_BODY',
  AUTH_VERIFY_MALFORMED_CLOUDCARD_RESPONSE_BODY = 'AUTH_VERIFY_MALFORMED_CLOUDCARD_RESPONSE_BODY',
  REFAUTH_PROVISION_REQUEST_FAIL = 'REFAUTH_PROVISION_REQUEST_FAIL',
  REFAUTH_PROVISION_CLIENT_FAIL_WITH_UNKNOWN_CODE = 'REFAUTH_PROVISION_CLIENT_FAIL_WITH_UNKNOWN_CODE',
  REFAUTH_PROVISION_CLIENT_FAIL_WITH_ALERT = 'REFAUTH_PROVISION_CLIENT_FAIL_WITH_ALERT',
  REFAUTH_PROVISION_CLIENT_FAIL_WITH_TECHNICAL_ERROR = 'REFAUTH_PROVISION_CLIENT_FAIL_WITH_TECHNICAL_ERROR',
  REFAUTH_PROVISION_CLIENT_FAIL_WITH_REQUEST_ERROR = 'REFAUTH_PROVISION_CLIENT_FAIL_WITH_REQUEST_ERROR',
  REFAUTH_PROVISION_CLIENT_FAIL_WITH_CONTRACT_EXECUTION_ERROR = 'REFAUTH_PROVISION_CLIENT_FAIL_WITH_CONTRACT_EXECUTION_ERROR',
  REFAUTH_PROVISION_FAIL_WITH_CARD_DATA_DECRYPTION_ERROR = 'REFAUTH_PROVISION_FAIL_WITH_CARD_DATA_DECRYPTION_ERROR',
  REFAUTH_PROVISION_FAIL_TRANSIENT_ERROR = 'REFAUTH_PROVISION_FAIL_TRANSIENT_ERROR',
  INVALID_CARD_PAN = 'INVALID_CARD_PAN',
  INVALID_EMAIL_ADDRESS = 'INVALID_EMAIL_ADDRESS',
  INVALID_PHONE_NUMBER = 'INVALID_PHONE_NUMBER',
  INVITATION_ALREADY_COMPLETED = 'INVITATION_ALREADY_COMPLETED',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  USER_ALREADY_EXIST = 'USER_ALREADY_EXIST',
  INVITATION_NOT_COMPLETED = 'INVITATION_NOT_COMPLETED',
  SECURITY_ASSERTION_SIGNATURE_VERIFICATION_FAILED = 'SECURITY_ASSERTION_SIGNATURE_VERIFICATION_FAILED',
  SECURITY_ASSERTION_INVALID_SIGNATURE = 'SECURITY_ASSERTION_INVALID_SIGNATURE',
}

export enum DefaultUiErrorMessages {
  AUTH_INIT_TOO_MANY_ACTIVE_AUTH_SESSIONS_FOR_USER = 'you have reached the maximum concurrent authentication context at the moment, please try again later',
  AUTH_INIT_UNKNOWN_USER = 'This user is not provisioned',
  AUTH_INIT_AUTHENTICATION_LOCKED = 'The authentication is locked, please wait',
  AUTH_INIT_AUTHENTICATION_FAILED = 'The authentication failed, please contact the authentication provider',
  AUTH_INIT_AUTHENTICATION_FAILED_SMS_PROVIDER_ERROR = 'The authentication failed (SMS), please contact the authentication provider',
  REFAUTH_PROVISION_CLIENT_FAIL_WITH_ALERT = 'Service contract possibly partially fulfilled. Messages intended for the applicant specify the situation',
  REFAUTH_PROVISION_CLIENT_FAIL_WITH_TECHNICAL_ERROR = 'A technical error occurred during the execution of the service contract',
  REFAUTH_PROVISION_CLIENT_FAIL_WITH_REQUEST_ERROR = 'Absence or inconsistency in the elements of the question',
  REFAUTH_PROVISION_CLIENT_FAIL_WITH_CONTRACT_EXECUTION_ERROR = 'Error when completing the service contract',
  REFAUTH_PROVISION_CLIENT_FAIL_WITH_UNKNOWN_CODE = 'Error occurred with unknown code',
  REFAUTH_PROVISION_FAIL_WITH_CARD_DATA_DECRYPTION_ERROR = 'Error occurred while decrypting card data',
  REFAUTH_PROVISION_FAIL_TRANSIENT_ERROR = 'Error occurred with provisioning request',
  INVALID_CARD_PAN = 'Error occurred with card provided',
  INVALID_EMAIL_ADDRESS = 'Error occurred with email provided',
  INVALID_PHONE_NUMBER = 'Error occurred with phone provided',
  INVITATION_ALREADY_COMPLETED = 'Error occurred with invitation provided',
  USER_NOT_FOUND = 'Error occurred with user id provided',
  USER_ALREADY_EXIST = 'Error occurred with user id provided',
  SECURITY_ASSERTION_SIGNATURE_VERIFICATION_FAILED = 'Error occurred when verifying response signature',
  SECURITY_ASSERTION_INVALID_SIGNATURE = 'Invalid Response Signature',
}